Update patch set 1
Patch Set 1:

Couple of comments on this merge:

I am aware this looks like a lot. 
But it is to merge the msgrtr repo into the messageService repo, and combine them as one.

There is alot of pre-existing technical debt in the msgrtr repo, and will have to be actioned at some point.
However, this story consisted of merging the repos and maintaining functionality. 

The csits run successfully against this change, however, the CSITs do not cover the whole code base.

All the unit tests pass successfully. But again, I am reluctant to depend on them due to the aforementioned technical debt (due to previously commented out tests, and other deficiencies I noticed while merging)

Patch-set: 1